+++ deser/src/de/impls.rs
@@ -2,7 +2,7 @@ use std::borrow::Cow;
 use std::collections::{BTreeMap, HashMap};
 use std::hash::BuildHasher;
 use std::hash::Hash;
-use std::mem::take;
+use std::mem::{take, MaybeUninit};
 
 use crate::de::{Deserialize, DeserializerState, MapSink, SeqSink, Sink, SinkHandle};
 use crate::descriptors::{Descriptor, NamedDescriptor, UnorderedNamedDescriptor};
@@ -119,11 +119,15 @@ impl Deserialize for u8 {
         SlotWrapper::make_handle(out)
     }
 
-    fn __private_byte_slice_to_vec(bytes: &[u8]) -> Option<Vec<Self>>
+    unsafe fn __private_slice_from_bytes(bytes: &[u8]) -> Option<Vec<Self>>
     where
         Self: Sized,
     {
         Some(bytes.to_vec())
+    }
+
+    fn __private_is_bytes() -> bool {
+        true
     }
 }
 
@@ -182,7 +186,7 @@ deserialize!(f64);
 
 impl<T: Deserialize> Sink for SlotWrapper<Vec<T>> {
     fn expecting(&self) -> Cow<'_, str> {
-        if T::__private_byte_slice_to_vec(&[][..]).is_some() {
+        if T::__private_is_bytes() {
             "bytes".into()
         } else {
             "vec".into()
@@ -191,8 +195,8 @@ impl<T: Deserialize> Sink for SlotWrapper<Vec<T>> {
 
     fn atom(&mut self, atom: Atom, _state: &DeserializerState) -> Result<(), Error> {
         match atom {
-            Atom::Bytes(value) => {
-                if let Some(bytes) = T::__private_byte_slice_to_vec(&value) {
+            Atom::Bytes(value) => unsafe {
+                if let Some(bytes) = T::__private_slice_from_bytes(&value) {
                     **self = Some(bytes);
                     Ok(())
                 } else {
@@ -201,7 +205,7 @@ impl<T: Deserialize> Sink for SlotWrapper<Vec<T>> {
                         format!("unexpected bytes, expected {}", self.expecting()),
                     ))
                 }
-            }
+            },
             other => Err(other.unexpected_error(&self.expecting())),
         }
     }
@@ -389,7 +393,7 @@ impl<'a, T: Deserialize> SeqSink for VecSink<'a, T> {
     fn descriptor(&self) -> &dyn Descriptor {
         static SLICE_DESCRIPTOR: NamedDescriptor = NamedDescriptor { name: "Vec" };
         static BYTES_DESCRIPTOR: NamedDescriptor = NamedDescriptor { name: "ByteVec" };
-        if T::__private_byte_slice_to_vec(&[]).is_some() {
+        if T::__private_is_bytes() {
             &BYTES_DESCRIPTOR
         } else {
             &SLICE_DESCRIPTOR
@@ -442,5 +446,194 @@ impl<'a> Sink for NullIgnoringSink<'a> {
 
     fn expecting(&self) -> Cow<'_, str> {
         "optional".into()
+    }
+}
+
+macro_rules! deserialize_for_tuple {
+    () => ();
+    ($($name:ident,)+) => (
+        impl<$($name: Deserialize),*> Deserialize for ($($name,)*) {
+            fn deserialize_into(out: &mut Option<Self>) -> SinkHandle {
+                SlotWrapper::make_handle(out)
+            }
+        }
+
+        impl<$($name: Deserialize),*> Sink for SlotWrapper<($($name,)*)> {
+            #[allow(non_snake_case)]
+            fn seq(&mut self, _state: &DeserializerState) -> Result<Box<dyn SeqSink + '_>, Error> {
+                struct TupleSink<'a, $($name,)*> {
+                    slot: &'a mut Option<($($name,)*)>,
+                    index: usize,
+                    $(
+                        $name: Option<$name>,
+                    )*
+                }
+
+                impl<'a, $($name: Deserialize,)*> SeqSink for TupleSink<'a, $($name,)*> {
+                    fn item(&mut self) -> Result<SinkHandle, Error> {
+                        let __index = self.index;
+                        self.index += 1;
+                        let mut __counter = 0;
+                        $(
+                            if __index == __counter {
+                                return Ok(Deserialize::deserialize_into(&mut self.$name));
+                            }
+                            __counter += 1;
+                        )*
+                        Err(Error::new(ErrorKind::WrongLength, "too many elements in tuple"))
+                    }
+
+                    fn finish(&mut self, _state: &DeserializerState) -> Result<(), Error> {
+                        *self.slot = Some(($(
+                            self.$name
+                                .take()
+                                .ok_or_else(|| Error::new(ErrorKind::WrongLength, "not enough elements in tuple"))?,
+                        )*));
+                        Ok(())
+                    }
+                }
+
+                Ok(Box::new(TupleSink {
+                    slot: self,
+                    index: 0,
+                    $(
+                        $name: None,
+                    )*
+                }))
+            }
+
+            fn expecting(&self) -> Cow<'_, str> {
+                "tuple".into()
+            }
+        }
+
+        deserialize_for_tuple_peel!($($name,)*);
+    )
+}
+
+macro_rules! deserialize_for_tuple_peel {
+    ($name:ident, $($other:ident,)*) => (deserialize_for_tuple!($($other,)*);)
+}
+
+deserialize_for_tuple! { T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, }
+
+impl<T: Deserialize, const N: usize> Deserialize for [T; N] {
+    fn deserialize_into(out: &mut Option<Self>) -> SinkHandle {
+        struct ArraySink<'a, T, const N: usize> {
+            out: &'a mut Option<[T; N]>,
+        }
+
+        impl<'a, T: Deserialize + 'a, const N: usize> Sink for ArraySink<'a, T, N> {
+            fn expecting(&self) -> Cow<'_, str> {
+                "array".into()
+            }
+
+            fn atom(&mut self, atom: Atom, _state: &DeserializerState) -> Result<(), Error> {
+                match atom {
+                    Atom::Bytes(value) => {
+                        if let Some(bytes) = unsafe { T::__private_slice_from_bytes(&value) } {
+                            if bytes.len() == N {
+                                *self.out = Some(unsafe {
+                                    let mut rv = MaybeUninit::<[T; N]>::uninit();
+                                    std::ptr::copy_nonoverlapping(
+                                        bytes.as_slice().as_ptr() as *const T,
+                                        rv.as_mut_ptr() as *mut T,
+                                        N,
+                                    );
+                                    rv.assume_init()
+                                });
+                                Ok(())
+                            } else {
+                                Err(Error::new(
+                                    ErrorKind::WrongLength,
+                                    "byte array of wrong length",
+                                ))
+                            }
+                        } else {
+                            Err(Error::new(
+                                ErrorKind::Unexpected,
+                                format!("unexpected bytes, expected {}", self.expecting()),
+                            ))
+                        }
+                    }
+                    other => Err(other.unexpected_error(&self.expecting())),
+                }
+            }
+
+            fn seq(&mut self, _state: &DeserializerState) -> Result<Box<dyn SeqSink + '_>, Error> {
+                Ok(Box::new(ArraySeqSink {
+                    slot: self.out,
+                    buffer: Some(unsafe { MaybeUninit::uninit().assume_init() }),
+                    element: None,
+                    index: 0,
+                }))
+            }
+        }
+
+        struct ArraySeqSink<'a, T, const N: usize> {
+            slot: &'a mut Option<[T; N]>,
+            buffer: Option<[MaybeUninit<T>; N]>,
+            element: Option<T>,
+            index: usize,
+        }
+
+        impl<'a, T, const N: usize> ArraySeqSink<'a, T, N> {
+            unsafe fn flush(&mut self) {
+                if let Some(element) = self.element.take() {
+                    let buffer = self.buffer.as_mut().unwrap();
+                    buffer[self.index].write(element);
+                    self.index += 1;
+                }
+            }
+        }
+
+        impl<'a, T, const N: usize> Drop for ArraySeqSink<'a, T, N> {
+            fn drop(&mut self) {
+                if std::mem::needs_drop::<T>() {
+                    if let Some(arr) = &mut self.buffer {
+                        for elem in &mut arr[0..self.index] {
+                            unsafe {
+                                std::ptr::drop_in_place(elem.as_mut_ptr());
+                            }
+                        }
+                    }
+                }
+            }
+        }
+
+        impl<'a, T: Deserialize, const N: usize> SeqSink for ArraySeqSink<'a, T, N> {
+            fn item(&mut self) -> Result<SinkHandle, Error> {
+                unsafe {
+                    self.flush();
+                }
+                if self.index >= N {
+                    Err(Error::new(
+                        ErrorKind::WrongLength,
+                        "too many elements in array",
+                    ))
+                } else {
+                    Ok(Deserialize::deserialize_into(&mut self.element))
+                }
+            }
+
+            fn finish(&mut self, _state: &DeserializerState) -> Result<(), Error> {
+                unsafe {
+                    self.flush();
+                }
+                if self.index != N {
+                    Err(Error::new(
+                        ErrorKind::WrongLength,
+                        "not enough elements in array",
+                    ))
+                } else {
+                    *self.slot = Some(unsafe {
+                        self.buffer.take().unwrap().as_ptr().cast::<[T; N]>().read()
+                    });
+                    Ok(())
+                }
+            }
+        }
+
+        SinkHandle::boxed(ArraySink { out })
     }
 }