+++ deser/tests/test_ser.rs
@@ -1,3 +1,5 @@
+use std::borrow::Cow;
+
 use deser::ser::for_each_event;
 use deser::{Atom, Event, Serialize};
 
@@ -18,4 +20,47 @@ fn test_optional() {
 
     let events = capture_events(&Some(42usize));
     assert_eq!(events, vec![Event::Atom(Atom::U64(42))]);
+}
+
+#[test]
+fn test_tuples() {
+    let events = capture_events(&(1, 2, 3));
+    assert_eq!(
+        events,
+        vec![
+            Event::SeqStart,
+            1i64.into(),
+            2i64.into(),
+            3i64.into(),
+            Event::SeqEnd,
+        ]
+    );
+}
+
+#[test]
+fn test_unit() {
+    let events = capture_events(&());
+    assert_eq!(events, vec![Event::Atom(Atom::Null)]);
+}
+
+#[test]
+fn test_array() {
+    let events = capture_events(&[1i64, 2, 3, 4]);
+    assert_eq!(
+        events,
+        vec![
+            Event::SeqStart,
+            1i64.into(),
+            2i64.into(),
+            3i64.into(),
+            4i64.into(),
+            Event::SeqEnd
+        ]
+    );
+
+    let events = capture_events(b"Hello");
+    assert_eq!(
+        events,
+        vec![Event::Atom(Atom::Bytes(Cow::Borrowed(b"Hello")))]
+    );
 }